WIP - Do a little refactoring
