+++ src/generation/object/lib/mod.rs
@@ -1,10 +1,12 @@
 mod cell;
+mod collapsed_cell;
 mod connection_type;
 mod failure;
 mod object_grid;
 mod object_name;
 
 pub use cell::Cell;
+pub use collapsed_cell::CollapsedCell;
 pub use connection_type::Connection;
 pub use failure::NoPossibleStatesFailure;
 pub use object_grid::ObjectGrid;
